Add a Jellybeans inspired theme (dark, colorful).
Based on the colors in the Vim theme here:
https://github.com/nanotech/jellybeans.vim

Yaml file based on the molokai one in the tree.

I've tried to use a fairly sensible palette to keep similar things in
similar color spectrums. Text is green, code is blue, markup is blue
green, etc. That said, I'm not an artist or a color person, so happy to
have suggestions on better arrangements here. I'm mostly trying to gain
some consistency in pallette across Vim and other tools.